Error fixes

+++ src/app/pages/register/register.page.ts
@@ -2,6 +2,7 @@ import { LoginService } from './../../services/login.service';
 import { Component, OnInit } from '@angular/core';
 import { FormGroup, FormBuilder } from "@angular/forms";
 import { Router } from '@angular/router';
+import { ToasterService } from 'src/app/services/toaster.service';
 
 @Component({
   selector: 'app-register',
@@ -9,11 +10,13 @@ import { Router } from '@angular/router';
   styleUrls: ['./register.page.scss'],
 })
 export class RegisterPage implements OnInit {
+  expression: RegExp = /^[A-Z0-9._%+-]+@[A-Z0-9.-]+\.[A-Z]{2,}$/i;
 
   registerForm!: FormGroup
   constructor(private loginService: LoginService,
     public formBuilder: FormBuilder,    
-    private router: Router) { }
+    private router: Router,
+    public toastr: ToasterService) { }
 
   ngOnInit() {
     this.createForm();
@@ -29,14 +32,18 @@ export class RegisterPage implements OnInit {
   }
 
   onSubmit() {
+    if(this.expression.test(this.registerForm.value.email)){
       this.loginService.createUser({...this.registerForm.value})
       .then(() => {
         this.registerForm.reset();
         localStorage.clear();
         this.router.navigate(['login']);
       }).catch((err) => {
-        console.log(err)
+        
       });
+    } else {
+      this.toastr.presentToast('Enter valid email', 'warning');
+    }
   }
